Create independent nodes.py
Determining if independent nodes exist in a graph using k-clique decision problem. Reducing a problem to a NP-hard problem which makes that problem a NP-hard problem and in this case as "independent nodes" is an NP problem as well (certificate and verification algorithm can be found), it is a NP-complete problem.