Create from clause to color
In the reduction from 3-SAT to 3-colorability, we convert 3-SAT problem with x variable and y clauses into a graph with n nodes (verices) and m edges. 
considerin a gadget with 6 nodes and 7 edges, the relationship between the number of nodes and edges with the number of variables and clauses is as follow.